confirm modal

+++ src/app/modules/gas-station/list-station/list-station.component.ts
@@ -1,8 +1,11 @@
-import { ChangeDetectorRef, Component, OnInit, ViewChild } from '@angular/core';
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { Router } from '@angular/router';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { debounceTime, takeUntil } from 'rxjs/operators';
+import { ConfirmDeleteComponent } from 'src/app/shared/components/confirm-delete/confirm-delete.component';
 import { LIST_STATUS } from 'src/app/shared/data-enum/list-status';
+import { IConfirmModalData } from 'src/app/shared/models/confirm-delete.interface';
 import { DestroyService } from 'src/app/shared/services/destroy.service';
 import { FilterService } from 'src/app/shared/services/filter.service';
 import { SortService } from 'src/app/shared/services/sort.service';
@@ -32,7 +35,8 @@ export class ListStationComponent implements OnInit {
     private destroy$: DestroyService,
     private cdr: ChangeDetectorRef,
     private router: Router,
-    private gasStationService: GasStationService
+    private gasStationService: GasStationService,
+    private modalService: NgbModal
   ) {
     this.dataSource = this.dataSourceTemp = [];
     this.sorting = sortService.sorting;
@@ -82,6 +86,18 @@ export class ListStationComponent implements OnInit {
   }
 
   deleteStation(stationId: string) {
-    this.gasStationService.deleteStation(stationId).subscribe(() => this.getListStation());
+    const modalRef = this.modalService.open(ConfirmDeleteComponent);
+    const data: IConfirmModalData = {
+      title: 'Xác nhận',
+      message: 'Bạn có chắc chắn muốn xoá  trạm 1 ?',
+      button: { class: 'btn-primary', title: 'Xác nhận' }
+    };
+    modalRef.componentInstance.data = data;
+
+    modalRef.result.then((result) => {
+      if (result) {
+        this.gasStationService.deleteStation(stationId).subscribe(() => this.getListStation());
+      }
+    });
   }
 }
